services and handlers

+++ app/models/model_user.go
@@ -38,8 +38,8 @@ func (s *User) SetCreatedAt(v time.Time) { s.CreatedAt = v }
 func (s *User) SetUpdatedAt(v time.Time) { s.UpdatedAt = v }
 
 type UserToken struct {
-	Id               int64      `db:"id"`
-	UserID           int64      `db:"user_id"`
+	Id               uint       `db:"id"`
+	UserID           uint       `db:"user_id"`
 	Device           string     `db:"device"`
 	RefreshToken     string     `db:"refresh_token"`
 	RefreshExpiresAt time.Time  `db:"refresh_expires_at"`
@@ -49,10 +49,10 @@ type UserToken struct {
 }
 
 // GetID - returns the id.
-func (s *UserToken) GetID() int64 { return s.Id }
+func (s *UserToken) GetID() uint { return s.Id }
 
 // SetID - sets the id.
-func (s *UserToken) SetID(id int64) { s.Id = id }
+func (s *UserToken) SetID(id uint) { s.Id = id }
 
 // IsExpired - check is token expired
 func (s *UserToken) IsExpired() bool {
